requirements.txt
django-bootstrap4
django-import-export
django-debug-toolbar
psycopg2
